Disable keyword style, so that identifiers like :comparator: are accepted
In older CHICKENs, keywords could be bound like regular symbols, which
was a bit of a strange behaviour.  In 5.0.2 and later this no longer
works, so we'll need to either quote the symbols using pipe characters
or disable the keyword style reader.  The latter is best for portable
code as it requires no actual changes to be made to the upstream code.